viventium_v0_4/telegram-viventium/TelegramVivBot/pyproject.toml
[project]
name = "chatgpt-telegram-bot"
version = "1.0.17"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chardet>=5.2.0",
    "duckduckgo-search==5.3.1",
    "fake-useragent>=2.2.0",
    "fastapi>=0.116.1",
    "fastmcp>=2.12.4",
    "google-api-python-client>=2.184.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "html2text>=2025.4.15",
    "httpx-socks>=0.10.1",
    "httpx[http2]>=0.28.1",
    "lxml-html-clean>=0.4.2",
    "mcp>=1.17.0",
    "msgspec>=0.19.0",
    "pdfminer-six==20240706",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[job-queue,rate-limiter,webhooks]>=22.3",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "pywhispercpp>=1.0.0",
    "livekit>=0.18.0",
    "livekit-api>=0.6.0",
]
